✨ Prevent Multi instances, "now playing", Taskbar Actions
Possible there's a fix for some Keybind issues with Accented Characters... Blame JS for being a bit stupid with Processing it as a 1-1 character with no way of telling them apart.
Possible fix for a random issue with the Queue I had where it something was missing and just entirely prevented usage of the app...
And all things mentioned above.